requirements.txt
numpy>=1.26,<2.0
pandas>=2.0,<3.0
scipy>=1.10
scikit-learn>=1.3
lightgbm>=4.6
torch>=2.0
tqdm>=4.66
PyYAML>=6.0
catboost>=1.2
xgboost>=2.1
